[BID-79] refactor(io): separate polars/pandas reader paths + sparse correctness
Checkpoint of the dual-path review work (Phases 0/1/3). Default behaviour on
clean input is bit-identical, verified by new reader parity tests; the pandas
transform path and the dense split builder remain, to be removed at cutover.

Phase 0 (oracle): value-level reader parity tests for Sage/DIA-NN/FragPipe/DELPI
(pandas path as reference) + a shared _parity_helpers; rewrite the split_tmt
sparse tests onto hand-built block-diagonal references so they survive deletion
of the dense builder; add read_diann(sparse=True) and _has_quant_values coverage.

Phase 1 (reader separation + divergence fixes): isolate every pandas transform in
a delete-at-cutover module (_readers_pandas.py), selected by a single explicit
ReaderEngine flag resolved once per read instead of scattered _is_polars(df)
sniffing. Fix the polars null-handling / parser-default divergences the pandas
path was masking: Sage scannr (raise on non-scan= token), leading-zero scan, and
null label; DELPI/MaxQuant null index keys; FragPipe null Spectrum; empty and
optional columns (null_values, infer_schema_length, diagonal_relaxed concat).

Phase 3 (sparse correctness + memory): split_tmt no longer 0-fills structurally
absent cells (dense_block instead of toarray/to_df) and raises on an incomplete
set map; _has_quant_values reports an all-absent sparse matrix as empty; .X is
unified to float32; the TMT to_peptide purity filter is applied sparse-natively
(drop CSC feature columns) so the block-diagonal is never densified.

+++ msmu/_core/_blockdiag.py
@@ -170,7 +170,10 @@ def to_dense_df(adata, layer: str | None = None) -> pd.DataFrame:
     """
     matrix = adata.X if layer is None else adata.layers[layer]
     if sp.issparse(matrix):
-        values = dense_block(matrix).astype(matrix.dtype, copy=False)
+        # dense_block already returns float64 with NaN for absent cells; keep that (do not re-cast to
+        # matrix.dtype) so the sparse and dense branches return the same dtype, and so a non-float .X
+        # cannot turn the NaN fill back into 0.
+        values = dense_block(matrix)
     else:
         values = np.asarray(matrix, dtype=float)
     return pd.DataFrame(values, index=adata.obs_names, columns=adata.var_names)
